pyproject.toml
[tool.poetry]
name = "ntfs-with-modes"
version = "0.1.0"
description = "Mount ntfs filesystems with linux file modes"
authors = ["Michael Lipschultz <michael.lipschultz@gmail.com>"]
readme = "README.md"
packages = [{include = "ntfs_with_modes"}]

[tool.poetry.dependencies]
python = "^3.7"
fusepy = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
